+++ .ghcid
@@ -12,3 +12,4 @@
 --restart=packages/htmx-lucid-contrib/htmx-lucid-contrib.cabal
 --restart=packages/nix/nix.cabal
 --restart=packages/nix-cache-server/nix-cache-server.cabal
+--restart=packages/servant-github-webhook-simple/servant-github-webhook-simple.cabal
