+++ packages/servant-github-webhook-simple/README.md
@@ -1,0 +1,3 @@
+# servant-github-webhook-simple
+
+An opinionated take on GitHub webhook handling for Servant, inspired by [servant-github-webhook](https://hackage.haskell.org/package/servant-github-webhook). Uses [crypton](https://hackage.haskell.org/package/crypton) for signature verification instead of the now archived [cryptonite](https://github.com/haskell-crypto/cryptonite).
