+++ packages/vira/vira.cabal
@@ -219,6 +219,7 @@ library
     , safecopy
     , servant
     , servant-event-stream
+    , servant-github-webhook-simple
     , servant-server
     , sqlite-simple
     , stm
@@ -348,6 +349,7 @@ executable vira
     , safecopy
     , servant
     , servant-event-stream
+    , servant-github-webhook-simple
     , servant-server
     , sqlite-simple
     , stm
@@ -482,6 +484,7 @@ executable vira-tests
     , safecopy
     , servant
     , servant-event-stream
+    , servant-github-webhook-simple
     , servant-server
     , sqlite-simple
     , stm
